requirements.txt
azure-storage-blob~=12.9.0
pandas~=1.4.1
adlfs~=2022.2.0
